Moving _flexible_scaler to _data
