+++ skcosmo/preprocessing/_data.py
